binance_bot/requirements.txt
python-binance==1.0.19
pandas==2.2.3
numpy==2.2.1
tabulate==0.9.0
pyyaml==6.0.2
typer==0.12.3
rich==13.7.1
